Allow save/load/clear.
In the course of this work, refined API for saving/loading--there's no
actual reason to have the core logic be interactive, since web vs. console
will have different IDataStorage implementations in the first place. Will
refine further.

It was also necessary to switch execute to a continuation-passing style,
but that should work fine both with React/Elmish mailbox dispatchers and
with console applications.