Add while loop support Game has approximately the right form now, but the actual shrek attack and bob attack are not working correctly, results in match case failure.
The problem seems to be that the API for native functions like Roll isn't well-thought-out.
Native is expecting a Const(Number), but it's actually getting an LocalRef("spec").

Okay, now that's kind of fixed (needs more thought), but the takeDamage event isn't able to
find "target" parameter because it didn't propagate from "attack" through "hit".