FSX version of XP tracking prototype, inspired by the book /Characteristics of Games/:
This implements a tracker to make a different XP design feasible,
in order to increase competitiveness between players while still
maintaining an overall cooperative flow. The new scheme is:

Instead of awarding XP to the whole party, XP is awarded on
a per-monster basis to all characters or creatures who damaged
or were damaged by that monster, proportional to the ratio
of damage taken + received by that player.

For example, a beholder is worth 10,000 XP. If Bob did 40 HP of
damage to the beholder, while Lara did 20 HP of damage and
took 30, and Harry took 10 HP of damage from it as well, then
Bob gets 40/100*10000 = 4000 XP, while Lara gets 5000 XP and
Harry gets 1000.

This is intended to somewhat discourage summoners and encourage
a bit of politicking among players, and a bit of recklessness.

It will also provide the basis for tracking damage ratios and
total damage stats between different PCs, enabling a different
form of campaign-like competition.