Make the sidebar a reasonable, scrollable size.
I don't love that we had to do so much 95vh instead of 100%--it feels awkward and ugly. But it's functional for now.

Note oddity: if we give table a width of 100% and a margin-right of 5em, the table overflows its parent div.
I thought border-box was supposed to prevent that but... something to be aware of as a possibility if you
see unexpected overflows. A workaround is to set padding-right instead.