Fix warnings, and experiment with serialization. (Needed for saving games AND for transmitting over WebRTC.)
One way of doing serialization--but notice how we've got a map specific to id=1 and we're not directly using it.
Maybe we should just directly serialize the map instead of pulling values out via get/tryGet?
But that might cause issues with inherited properties when we e.g. let Bob have a sword, with its own id.