(still 1 red) Implemented budgeted, but there's a test failure still.
proto1 is failing for reasons due to the label shown by either when a
child is unselected. We need to distinguish between the unselected label
(which should be the lowest label, e.g. Acrobatics +1) and the label
when selected (which could ranged from Acrobatics +1 to Acrobatics +3
or whatever). Either needs to access it somehow.