Tentative partial fix for issue #52 w/rt Sword and Dagger. Feels kind of messy--is using label as a signal of whether to collapse AND really the right choice?
Also, I don't love the UX. Having an unselect checkbox (only) next to Sword and not Main-gauche feels kind of unintuitive. How is the user supposed to understand
that it's possible to make Main-gauche go away? I think it needs to become a Checked tied to the key for Sword and Dagger, and not an unconditional.

Proposed rule: a 'full' AND can have its headers suppressed, but all of its unconditional leaves need to be converted to Checked at that point, probably via
manipulating the renderAPI passed down.